BugFix #53
Fixes #53.
Modified the way the JSON is sent. Header added to allow correct detection of a valid JSON object.

During a weak signal event, malformed or lost packets containing JSON objects will cause the service to close the session with the RemoteUser, instead of crashing the service and forcing FailSafe. 

Fixes #53